requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
mysql-connector==2.1.6
numpy==1.15.3
requests==2.20.0
urllib3==1.24
Werkzeug==0.14.1
